Replace old setup.py with setupmono.py, modified to work on both linux and windows. Remove old unused Makefiles and out of date README.txt.
