+++ pythonnet/setup.py
@@ -1,61 +1,199 @@
-#!/usr/bin/env python
-# ==========================================================================
-# This software is subject to the provisions of the Zope Public License,
-# Version 2.1 (ZPL).  A copy of the ZPL should accompany this distribution.
-# THIS SOFTWARE IS PROVIDED "AS IS" AND ANY AND ALL EXPRESS OR IMPLIED
-# WARRANTIES ARE DISCLAIMED, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-# WARRANTIES OF TITLE, MERCHANTABILITY, AGAINST INFRINGEMENT, AND FITNESS
-# FOR A PARTICULAR PURPOSE.
-# ==========================================================================
-"""Setup file for Mono clr.so
-
-Author: Christian Heimes <christian(at)cheimes(dot)de>
 """
-
-import os
+Setup script for building clr.pyd and dependencies using mono and into
+an egg or wheel.
+"""
+from setuptools import setup, Extension
+from distutils.command.build_ext import build_ext
+from distutils.command.build_scripts import build_scripts
+from distutils.command.install_lib import install_lib
+from distutils.sysconfig import get_config_vars
+from platform import architecture
+from subprocess import Popen, CalledProcessError, PIPE, check_call
+from glob import glob
+import shutil
 import sys
-from distutils.core import setup
-from distutils.core import Extension
-import subprocess
+import os
 
-VERSION = "%i.%i" % sys.version_info[:2] 
+CONFIG = "Release" # Release or Debug
+DEVTOOLS = "MsDev" if sys.platform == "win32" else "Mono"
+VERBOSITY = "minimal" # quiet, minimal, normal, detailed, diagnostic
 
-def pkgconfig(*packages, **kw):
-    """From http://aspn.activestate.com/ASPN/Cookbook/Python/Recipe/502261
-    XXX cmd = "export PKG_CONFIG_PATH=/opt/mono-2.8/lib/pkgconfig:... early 2011
+if DEVTOOLS == "MsDev":
+    from distutils import msvc9compiler
+    msvc9compiler.VERSION = 11
+
+    cc = msvc9compiler.MSVCCompiler()
+    cc.initialize()
+    _xbuild = cc.find_exe("msbuild.exe")
+    _defines_sep = ";"
+    _config = "%sWin" % CONFIG
+    _npython_exe = "nPython.exe"
+
+elif DEVTOOLS == "Mono":
+    _xbuild = "xbuild"
+    _defines_sep = ","
+    _config = "%sMono" % CONFIG
+    _npython_exe = "npython"
+
+else:
+    raise NotImplementedError("DevTools %s not supported (use MsDev or Mono)" % DEVTOOLS)
+
+_platform = "x64" if architecture()[0] == "64bit" else "x86"
+
+class PythonNET_BuildExt(build_ext):
+
+    def build_extension(self, ext):
+        """
+        Builds the .pyd file using msbuild or xbuild.
+        """
+        if ext.name != "clr":
+            return build_ext.build_extension(self, ext)
+
+        dest_file = self.get_ext_fullpath(ext.name)
+        dest_dir = os.path.dirname(dest_file)
+        if not os.path.exists(dest_dir):
+            os.makedirs(dest_dir)
+
+        defines = [
+            "PYTHON%d%s" % (sys.version_info[:2]),
+            "UCS2" if sys.maxunicode < 0x10FFFF else "UCS4",
+        ]
+
+        if CONFIG == "Debug":
+            defines.extend(["DEBUG", "TRACE"])
+
+        cmd = [
+            _xbuild,
+            "pythonnet.sln",
+            "/p:Configuration=%s" % _config,
+            "/p:Platform=%s" % _platform,
+            "/p:DefineConstants=\"%s\"" % _defines_sep.join(defines),
+            "/p:PythonBuildDir=%s" % os.path.abspath(dest_dir),
+            "/p:NoNuGet=true",
+            "/verbosity:%s" % VERBOSITY,
+        ]
+
+        self.announce("Building: %s" % " ".join(cmd))
+        check_call(" ".join(cmd) + " /t:Clean", shell=True)
+        check_call(" ".join(cmd) + " /t:Build", shell=True)
+
+        if DEVTOOLS == "Mono":
+            self._build_monoclr(ext)
+
+
+    def _build_monoclr(self, ext):
+        mono_libs = _check_output("pkg-config --libs mono-2", shell=True)
+        mono_cflags = _check_output("pkg-config --cflags mono-2", shell=True)
+        glib_libs = _check_output("pkg-config --libs glib-2.0", shell=True)
+        glib_cflags = _check_output("pkg-config --cflags glib-2.0", shell=True)
+        cflags = mono_cflags.strip() + " " + glib_cflags.strip()
+        libs = mono_libs.strip() + " " + glib_libs.strip()
+
+        # build the clr python module
+        clr_ext = Extension("clr",
+                    sources=[
+                        "src/monoclr/pynetinit.c",
+                        "src/monoclr/clrmod.c"
+                    ],
+                    extra_compile_args=cflags.split(" "),
+                    extra_link_args=libs.split(" "))
+
+        build_ext.build_extension(self, clr_ext)
+
+        # build the clr python executable
+        sources = [
+            "src/monoclr/pynetinit.c",
+            "src/monoclr/python.c",
+        ]
+
+        macros = ext.define_macros[:]
+        for undef in ext.undef_macros:
+            macros.append((undef,))
+
+        objects = self.compiler.compile(sources,
+                                        output_dir=self.build_temp,
+                                        macros=macros,
+                                        include_dirs=ext.include_dirs,
+                                        debug=self.debug,
+                                        extra_postargs=cflags.split(" "),
+                                        depends=ext.depends)
+
+        output_dir = os.path.dirname(self.get_ext_fullpath(ext.name))
+        py_libs = get_config_vars("BLDLIBRARY")[0]
+        libs += " " + py_libs
+
+        self.compiler.link_executable(objects,
+                                      _npython_exe,
+                                      output_dir=output_dir,
+                                      libraries=self.get_libraries(ext),
+                                      library_dirs=ext.library_dirs,
+                                      runtime_library_dirs=ext.runtime_library_dirs,
+                                      extra_postargs=libs.split(" "),
+                                      debug=self.debug)
+
+
+class PythonNET_InstallLib(install_lib):
+
+    def install(self):
+        if not os.path.isdir(self.build_dir):
+            self.warn("'%s' does not exist -- no Python modules to install" %
+                        self.build_dir)
+            return
+
+        if not os.path.exists(self.install_dir):
+            self.mkpath(self.install_dir)
+
+        # only copy clr.pyd and its dependencies
+        for pattern in ("clr.*", "Python.Runtime.*"):
+            for srcfile in glob(os.path.join(self.build_dir, pattern)):
+                destfile = os.path.join(self.install_dir, os.path.basename(srcfile))
+                self.copy_file(srcfile, destfile)
+    
+
+class PythonNET_BuildScripts(build_scripts):
+
+    def finalize_options(self):
+        build_scripts.finalize_options(self)
+
+        # fixup scripts to look in the build_ext output folder
+        if self.scripts:
+            build_ext = self.get_finalized_command("build_ext")
+            output_dir = os.path.dirname(build_ext.get_ext_fullpath(_npython_exe))
+            scripts = []
+            for script in self.scripts:
+                if os.path.exists(os.path.join(output_dir, script)):
+                    script = os.path.join(output_dir, script)
+                scripts.append(script)
+            self.scripts = scripts
+
+
+def _check_output(*popenargs, **kwargs):
+    """subprocess.check_output from python 2.7.
+    Added here to support building for earlier versions
+    of Python.
     """
-    flag_map = {'-I': 'include_dirs', '-L': 'library_dirs', '-l': 'libraries'}
-    cmd = "export PKG_CONFIG_PATH=/lib/pkgconfig; pkg-config --libs --cflags %s" % ' '.join(packages)
-    popen = subprocess.Popen(cmd, shell=True, close_fds=True, stdout=subprocess.PIPE)
-    popen.wait()
-    if popen.returncode != 0:
-        raise RuntimeError("An error has occured")
-    output = popen.stdout.read().strip()
+    process = Popen(stdout=PIPE, *popenargs, **kwargs)
+    output, unused_err = process.communicate()
+    retcode = process.poll()
+    if retcode:
+        cmd = kwargs.get("args")
+        if cmd is None:
+            cmd = popenargs[0]
+        raise CalledProcessError(retcode, cmd, output=output)
+    return output
 
-    for token in output.split():
-        if flag_map.has_key(token[:2]):
-            kw.setdefault(flag_map.get(token[:2]), []).append(token[2:])
-        else: # throw others to extra_link_args
-            kw.setdefault('extra_link_args', []).append(token)
 
-    for k, v in kw.iteritems(): # remove duplicated
-        kw[k] = list(set(v))
-
-    return kw
-#argsDict['include_dirs'] += ['/opt/mono-2.8/include','/include:usr/include','/usr/include/glib-2.0','/usr/lib/glib-2.0/include']
-argsDict = pkgconfig('glib-2.0', 'mono-2')
-
-clr = Extension('clr',
-    ['src/monoclr/clrmod.c', 'src/monoclr/pynetinit.c'],
-    depends=['src/monoclr/pynetclr.h'],
-    **argsDict
+if __name__ == "__main__":
+    setup(name="pythonnet",
+        ext_modules=[
+            Extension("clr", sources=[])
+        ],
+        scripts=[_npython_exe],
+        zip_safe=False,
+        cmdclass={
+            "build_ext" : PythonNET_BuildExt,
+            "build_scripts" : PythonNET_BuildScripts,
+            "install_lib" : PythonNET_InstallLib
+        }
     )
 
-extensions = []
-if os.name == "posix":
-    extensions.append(clr)
-
-setup(name="clr",
-    ext_modules = extensions,
-    scripts = ["src/monoclr/clrpython%s" % VERSION],
-    )
